+++ .github/workflows/deploy.yaml
@@ -32,14 +32,14 @@ jobs:
       - name: Set up Node
         uses: actions/setup-node@1d0ff469b7ec7b3cb9d8673fde0c81c44821de2a # v4.2.0
         with:
-          node-version: 22
-          cache: npm
+          node-version: lts/*
+          cache: yarn
 
       - name: Install packages
-        run: npm ci
+        run: yarn install --immutable
 
       - name: Build
-        run: npm run build
+        run: yarn run build
 
       # https://github.com/actions/configure-pages
       - name: Setup Pages
