+++ .pre-commit-config.yaml
@@ -12,7 +12,7 @@ repos:
     rev: v5.0.0
     hooks:
       - id: check-added-large-files  # Prevents giant files from being committed
-      - id: check-json               # Check that JSON files are valid
+      # - id: check-json               # Check that JSON files are valid
       - id: check-merge-conflict     # Check merge conflict strings
       - id: check-symlinks           # Ensure symlinks have a valid target
       - id: check-toml               # Ensure toml files are valid
@@ -46,3 +46,9 @@ repos:
     rev: "39.171.1"
     hooks:
       - id: renovate-config-validator
+
+  - repo: https://gitlab.com/bmares/check-json5
+    rev: "v1.0.0"
+    hooks:
+      # Check that JSON files are valid JSON5
+      - id: check-json5
