+++ package.json
@@ -4,33 +4,33 @@
   "version": "0.0.0",
   "type": "module",
   "scripts": {
-    "dev": "vite --host",
     "build": "tsc -b && vite build",
-    "lint": "eslint . && tsc --noEmit",
+    "coverage": "vitest run --coverage",
+    "dev": "vite --host",
+    "lint": "eslint .",
     "preview": "vite preview",
-    "test": "vitest run"
+    "test": "vitest"
   },
   "dependencies": {
-    "react": "19.0.0",
-    "react-dom": "19.0.0",
-    "react-router": "7.2.0",
-    "tailwindcss": "4.0.8"
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
+    "react-router": "^7.2.0"
   },
   "devDependencies": {
-    "@eslint/js": "9.21.0",
-    "@tailwindcss/vite": "4.0.8",
-    "@types/react": "19.0.10",
-    "@types/react-dom": "19.0.4",
-    "@vitejs/plugin-react": "4.3.4",
-    "eslint": "9.21.0",
-    "eslint-plugin-react": "7.37.4",
-    "eslint-plugin-react-hooks": "5.1.0",
-    "eslint-plugin-react-refresh": "0.4.19",
-    "eslint-plugin-tsdoc": "0.4.0",
-    "globals": "16.0.0",
-    "typescript": "5.7.3",
-    "typescript-eslint": "8.24.1",
-    "vite": "6.1.1",
-    "vitest": "3.0.6"
+    "@eslint/js": "^9.19.0",
+    "@tailwindcss/vite": "^4.0.8",
+    "@types/react": "^19.0.8",
+    "@types/react-dom": "^19.0.3",
+    "@vitejs/plugin-react": "^4.3.4",
+    "eslint": "^9.19.0",
+    "eslint-plugin-react": "^7.37.4",
+    "eslint-plugin-react-hooks": "^5.1.0",
+    "eslint-plugin-react-refresh": "^0.4.18",
+    "globals": "^15.14.0",
+    "tailwindcss": "^4.0.8",
+    "typescript": "~5.7.2",
+    "typescript-eslint": "^8.22.0",
+    "vite": "^6.1.0",
+    "vitest": "^3.0.6"
   }
 }
