+++ src/components/GameCell.tsx
